revert SessionMain refactor (now tracked on feature/session-main-split)
