+++ src/cpp/session/CMakeLists.txt
@@ -58,15 +58,9 @@ set (SESSION_SOURCE_FILES
    SessionClientEvent.cpp
    SessionClientEventQueue.cpp
    SessionClientEventService.cpp
-   SessionClientInit.cpp
-   SessionConsoleInput.cpp
    SessionConsoleProcess.cpp
    SessionContentUrls.cpp
-   SessionDirs.cpp
-   SessionFork.cpp
-   SessionRpc.cpp
-   SessionHttpMethods.cpp
-   SessionInit.cpp
+   SessionSSH.cpp
    SessionMain.cpp
    SessionMainOverlay.cpp
    SessionModuleContext.cpp
@@ -75,11 +69,8 @@ set (SESSION_SOURCE_FILES
    SessionPersistentState.cpp
    SessionPostback.cpp
    SessionRUtil.cpp
-   SessionSSH.cpp
    SessionSourceDatabase.cpp
    SessionSourceDatabaseSupervisor.cpp
-   SessionSuspend.cpp
-   SessionUriHandlers.cpp
    SessionUserSettings.cpp
    SessionWorkerContext.cpp
    http/SessionHttpConnectionQueue.cpp
