--- src/cpp/session/SessionFork.hpp
@@ -1,33 +1,0 @@
-/*
- * SessionFork.hpp
- *
- * Copyright (C) 2009-16 by RStudio, Inc.
- *
- * Unless you have received this program directly from RStudio pursuant
- * to the terms of a commercial license agreement with RStudio, then
- * this program is licensed to you under the terms of version 3 of the
- * GNU Affero General Public License. This program is distributed WITHOUT
- * ANY EXPRESS OR IMPLIED WARRANTY, INCLUDING THOSE OF NON-INFRINGEMENT,
- * MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE. Please refer to the
- * AGPL (http://www.gnu.org/licenses/agpl-3.0.txt) for more details.
- *
- */
-
-#ifndef SESSION_FORK_HPP
-#define SESSION_FORK_HPP
-
-namespace rstudio {
-namespace session {
-namespace fork {
-
-void setupForkHandlers();
-bool wasForked();
-bool haveRunningChildren();
-void initThreadId();
-
-} // namespace fork
-} // namespace session
-} // namespace rstudio
-
-#endif
-
